Refactor FSD entities
