+++ src/lib/entities/podcast/index.ts
@@ -1,2 +1,3 @@
 export { default as PodcastTile } from './ui/podcast-tile.svelte';
 export { default as PodcastDisplay } from './ui/podcast-display.svelte';
+export { podcastList, podcastGet } from './api/requests';
