+++ src/lib/shared/api/types.d.ts
@@ -1,22 +1,3 @@
-export interface SBFile {
-  name: string;
-}
-
-export interface SBPodcast {
-  id: number;
-  title: string;
-  author: string;
-  episodes: SBEpisode[];
-  coverArt: SBFile;
-}
-
-export interface SBEpisode {
-  id: number;
-  title: string;
-  duration: number;
-  audio: SBFile;
-}
-
 export interface Episode {
   id: number;
   title: string;
